03.02 commit

+++ kodilla-collections-advanced/src/main/java/com/kodilla/collections/adv/maps/homework/School.java
@@ -12,7 +12,7 @@ public class School {
         }
     }
 
-    public double getNumberofStudentsInSchool() {
+    public int getNumberofStudentsInSchool() {
         int sum = 0;
         for (int number : numberOfStudents)
             sum += number;
